pyproject.toml
[project]
name = "transolver-hf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Brian Staber", email = "brian.staber@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "huggingface-hub>=0.36.0",
    "mlflow>=3.6.0",
    "nvidia-physicsnemo>=1.3.0",
    "pyplaid>=0.1.10",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
